Revert "revert: "fix(query-performance): speed up trends queries by reducing amount of data streamed between shards" (#13496)"
This reverts commit aadda18cda81f8b92bcb0116a91cca510536eb38.
